backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
python-docx
pdfplumber
